# gdtracker-go-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GDTracker Go API
  version: 0.5.0
  description: |
    Go backend for GDTracker. Session auth and CSRF mirror the Spring API contract (see `docs/AUTH.md`).
    Routes mirror `gdtracker-api` for games, categories, tags, tasks, features, game configuration, archives, game exceptions,
    ingest plane (players, integration ping), game events (definitions + ingest), traces (heatmap), and feedback.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/games/{gameId}/features:
    get:
      tags: [Features]
      operationId: listFeatures
      parameters:
        - $ref: "#/components/parameters/GameId"
        - name: archived
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Feature" }
        "401": { description: Not authenticated }
        "404": { description: Game not found or not owned }
    post:
      tags: [Features]
      operationId: createFeature
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeatureUpsertRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Feature" }
        "400": { description: Validation error }
        "401": { description: Not authenticated }
        "404": { description: Game not found or not owned }
        "409": { description: Duplicate feature name }
  /api/games/{gameId}/features/task-progress:
    get:
      tags: [Features]
      operationId: listFeatureTaskProgress
      parameters:
        - $ref: "#/components/parameters/GameId"
        - name: archived
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FeatureTaskProgressRow" }
        "401": { description: Not authenticated }
        "404": { description: Game not found or not owned }
  /api/games/{gameId}/features/{id}:
    put:
      tags: [Features]
      operationId: updateFeature
      parameters:
        - $ref: "#/components/parameters/GameId"
        - $ref: "#/components/parameters/FeatureId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeatureUpsertRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Feature" }
        "400": { description: Validation error }
        "401": { description: Not authenticated }
        "404": { description: Feature not found }
        "409": { description: Conflict (archived feature or duplicate name) }
    delete:
      tags: [Features]
      operationId: deleteFeature
      parameters:
        - $ref: "#/components/parameters/GameId"
        - $ref: "#/components/parameters/FeatureId"
      responses:
        "204": { description: Deleted }
        "401": { description: Not authenticated }
        "404": { description: Feature not found }
        "409": { description: Has subfeatures or tasks }
  /api/games/{gameId}/features/{id}/archive:
    post:
      tags: [Features]
      operationId: archiveFeature
      parameters:
        - $ref: "#/components/parameters/GameId"
        - $ref: "#/components/parameters/FeatureId"
        - name: includeSummary
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Summary when includeSummary=true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "204": { description: No content }
        "401": { description: Not authenticated }
        "404": { description: Feature not found }
  /api/games/{gameId}/features/{id}/unarchive:
    post:
      tags: [Features]
      operationId: unarchiveFeature
      parameters:
        - $ref: "#/components/parameters/GameId"
        - $ref: "#/components/parameters/FeatureId"
        - name: includeSummary
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Summary when includeSummary=true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "204": { description: No content }
        "401": { description: Not authenticated }
        "404": { description: Feature not found }
...
components:
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema: { type: string }
...
    FeatureId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ingest-token
  schemas:
...
    Feature:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        status: { type: string }
        color: { type: string }
        parentId: { type: string, nullable: true }
        archived: { type: boolean }
        archivedAt: { type: string, nullable: true, format: date-time }
    FeatureUpsertRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        status: { type: string, nullable: true }
        color: { type: string, nullable: true }
        parentId: { type: string, nullable: true }
    FeatureTaskProgressRow:
      type: object
      required: [featureId, totalDirect, doneDirect, rolledUpTotal, rolledUpDone]
      properties:
        featureId: { type: string }
        totalDirect: { type: integer, format: int64 }
        doneDirect: { type: integer, format: int64 }
        rolledUpTotal: { type: integer, format: int64 }
        rolledUpDone: { type: integer, format: int64 }
...